+++ tests/testthat/test_add_node.r
@@ -20,7 +20,29 @@ test_that("not a DAG", {
 })
 
 test_that("error when node name already in DAG", {
+  # error with standard node() call
   expect_error(empty_dag() + node("A", "rnorm") + node("A", "rbernoulli"))
+
+  # no error when node is time-varying
+  expect_no_error(empty_dag() + node("A", "rnorm") + node_td("A", rnorm))
+
+  # error when node is time-varying, but also of type "time_to_event" or
+  # "competing_events"
+  expect_error(empty_dag() + node("A", "rnorm") +
+                 node_td("A", "time_to_event", prob_fun=0.1))
+  expect_error(empty_dag() + node("A", "rnorm") +
+                 node_td("A", "competing_events", prob_fun=0.1))
+})
+
+test_that("error when DAG would become cyclic", {
+  expect_error({
+    dag <- empty_dag() +
+      node("X", type="rnorm") +
+      node("A", type="binomial", formula= ~ -2 + B*1) +
+      node("B", type="binomial", formula= ~ -3 + A*1)
+  }, paste0("Adding node 'B' as specified is impossible, ",
+            "because it would make the DAG cyclic through the path:\n",
+            "A -> B -> A"))
 })
 
 test_that("works with multiple names", {