+++ tests/testthat/test_as.igraph.DAG.r
@@ -7,4 +7,10 @@ test_that("general test case", {
     node("D", type="gaussian", parents=c("A", "B", "C"))
   g <- as.igraph(dag)
   expect_true(igraph::is_igraph(g))
+
+  # excluding root nodes / td nodes
+  g <- as.igraph(dag, include_root_nodes=FALSE, include_td_nodes=FALSE)
+  expect_true(igraph::is_igraph(g))
 })
+
+