+++ tests/testthat/test_input_checks_do.r
@@ -2,7 +2,7 @@
 dag <- empty_dag() +
   node("A", type="rnorm", mean=23414, sd=234) +
   node("B", type="rcategorical") +
-  node("C", type="binomial", parents=c("A", "C"))
+  node("C", type="binomial", parents=c("A", "B"))
 
 test_that("not a DAG", {
   expect_error(do("", names="A", values=1))